requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
APScheduler==3.6.3
async-timeout==4.0.2
asynctest==0.13.0
attrs==21.4.0
autopep8==1.6.0
backports.zoneinfo==0.2.1
cachetools==4.2.2
certifi==2022.6.15
charset-normalizer==2.0.12
dateparser==1.1.1
frozenlist==1.3.0
idna==3.3
multidict==6.0.2
psycopg2-binary==2.9.3
pyaes==1.6.1
pyasn1==0.4.8
pycodestyle==2.8.0
python-dateutil==2.8.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
regex==2022.3.2
requests==2.28.0
rsa==4.8
six==1.16.0
Telethon==1.24.0
toml==0.10.2
tornado==6.1
typing_extensions==4.2.0
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
yarl==1.7.2
